fix: dynamic import를 클라이언트 컴포넌트로 분리하여 metadata 충돌 해소
